BJS2-78641 1.Добавление Equals и HashCode в сущности Hash и Url 2. Создание интерфейса для UrlService
